@@ -1,7 +1,6 @@
 package sockettransport_test
 
 import (
-	"io/ioutil"
 	"net"
 	"os"
 	"path"
@@ -55,10 +54,10 @@ func TestTcp(t *testing.T) {
 func TestUnix(t *testing.T) {
 	_, require := makeAR(t)
 
-	tmpdir, e := ioutil.TempDir("", "sockettransport-test")
+	dir, e := os.MkdirTemp("", "sockettransport-test")
 	require.NoError(e)
-	defer os.RemoveAll(tmpdir)
-	addr := path.Join(tmpdir, "unix.sock")
+	defer os.RemoveAll(dir)
+	addr := path.Join(dir, "unix.sock")
 	listener, e := net.Listen("unix", addr)
 	require.NoError(e)
 	defer listener.Close()
